+++ Thunderstore/CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
 
+## 1.1.0 - 2026-07-5
+- Fixed dynamic NavMesh generation at the company building
+- Added some config options to control the dynamic regeneration of the Company Navmesh
+- Added more dropdown links to the Company NavMesh
+- Fixed some parts of the NavMesh being generated inside of inaccessible areas of the Company Building
+
 ## 1.0.0 - 2026-07-3
 - Initial release